Added Token exchange filter to replace openid token with added scope token
This is to implement stateless authorization. User is authenticated on
UI with Keycloak auth server and gets an access token with just openid
and email scope. Then when resource server APIs are called by ui app,
this access token is sent as a bearer token. The filter present at
Spring Cloud Gateway will replace this token with added scopes and then
send this access token as bearer token to resource servers. Resource
server will validate the signature of the token and then serve the
request.